problem 2 solved

+++ problem2/prime_time_test.go
@@ -5,15 +5,15 @@ import "testing"
 func Test_isPrime(t *testing.T) {
 	tests := []struct {
 		name string
-		n    int
+		n    float64
 		want bool
 	}{
-		{"one", 1, false},
-		{"two", 2, true},
-		{"three", 3, true},
-		{"four", 4, false},
-		{"five", 5, true},
-		{"seventeen", 17, true},
+		{"one", 1.0, false},
+		{"two", 2.0, true},
+		{"three", 3.0, true},
+		{"four", 4.0, false},
+		{"five", 5.0, true},
+		{"seventeen", 17.0, true},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
